refactor: listbox context 파일 생성 및 관심사 로직 분리
