+++ src/components/Listbox/listbox.tsx
@@ -1,33 +1,14 @@
-import React, {
-  Dispatch,
-  MutableRefObject,
-  SetStateAction,
-  createContext,
-  forwardRef,
-  useContext,
-  useRef,
-  useState,
-} from 'react';
+import React, { Dispatch, SetStateAction, forwardRef, useRef, useState } from 'react';
 import { selectContainer, selectButton, optionList, optionButton, optionItem } from './listbox.css';
 import { useOutsideClick } from '@/hooks/useOutsideClick';
 import { cx } from '@/utils/cx';
+import { ListboxContext, useListboxContext } from './context';
 
 // TODO: 타입 확장하기
 type Item = {
   value: NonNullable<string>;
   option: string;
 };
-
-type ListboxContextState = {
-  onSelectedOptionChange: Dispatch<SetStateAction<string>>;
-  selectedOption: string;
-  items: Item[];
-  isOpen: boolean;
-  setIsOpen: Dispatch<SetStateAction<boolean>>;
-  optionRefs: MutableRefObject<HTMLButtonElement[]>;
-};
-
-const ListboxContext = createContext<ListboxContextState | null>(null);
 
 type ListboxProps = {
   onSelectedOptionChange: Dispatch<SetStateAction<string>>;
@@ -39,13 +20,9 @@ type ListboxProps = {
 export const Listbox = forwardRef<HTMLDivElement, ListboxProps>(
   ({ onSelectedOptionChange, selectedOption, items, label }, ref) => {
     const [isOpen, setIsOpen] = useState(false);
-    const selectRef = useRef<HTMLDivElement>(null);
-    const optionRefs = useRef<HTMLButtonElement[]>([]);
-
-    useOutsideClick(selectRef, () => setIsOpen(false));
 
     return (
-      <ListboxContext.Provider value={{ onSelectedOptionChange, selectedOption, items, isOpen, setIsOpen, optionRefs }}>
+      <ListboxContext.Provider value={{ onSelectedOptionChange, selectedOption, items, isOpen, setIsOpen }}>
         <div ref={ref} className={selectContainer} role="listbox" aria-expanded={isOpen} aria-haspopup="listbox">
           <ListboxTrigger label={label} />
           {isOpen && <ListboxOptions />}
@@ -56,16 +33,16 @@ export const Listbox = forwardRef<HTMLDivElement, ListboxProps>(
 );
 
 const ListboxTrigger = ({ label }: { label: string }) => {
-  const context = useContext(ListboxContext);
-
-  if (!context) {
-    throw new Error('ListboxTrigger must be used within a Listbox');
-  }
-
+  const context = useListboxContext();
   const { selectedOption, setIsOpen, isOpen } = context;
+
+  const selectRef = useRef<HTMLButtonElement>(null);
+
+  useOutsideClick(selectRef, () => setIsOpen(false));
 
   return (
     <button
+      ref={selectRef}
       className={`${selectButton} ${isOpen ? 'on' : ''}`}
       onClick={() => setIsOpen(!isOpen)}
       aria-haspopup="listbox"
@@ -77,13 +54,10 @@ const ListboxTrigger = ({ label }: { label: string }) => {
 };
 
 const ListboxOptions = () => {
-  const context = useContext(ListboxContext);
+  const context = useListboxContext();
+  const { items, selectedOption, onSelectedOptionChange, setIsOpen } = context;
 
-  if (!context) {
-    throw new Error('ListboxOptions must be used within a Listbox');
-  }
-
-  const { items, selectedOption, onSelectedOptionChange, optionRefs, setIsOpen } = context;
+  const optionRefs = useRef<HTMLButtonElement[]>([]);
 
   const handleKeyDown = (event: React.KeyboardEvent<HTMLButtonElement>) => {
     if (event.key === 'ArrowDown' || (event.key === 'Tab' && !event.shiftKey)) {
